feat: Introduce new intake-related fields and indexes to the database, and refactor practice client intake handlers and webhook processing logic.
