--- src/modules/practice-client-intakes/handlers/index.ts
@@ -1,13 +1,0 @@
-/**
- * Practice Client Intakes Handlers Index
- *
- * Exports all practice client intake webhook handlers
- */
-
-export { handlePracticeClientIntakeSucceeded } from './succeeded';
-export { handlePracticeClientIntakeFailed } from './failed';
-export { handlePracticeClientIntakeCanceled } from './canceled';
-export {
-  findPracticeClientIntakeByPaymentIntent,
-  findPracticeClientIntakeByCheckoutSession,
-} from './helpers';
