+++ src/modules/practice-client-intakes/http.ts
@@ -1,183 +1,25 @@
-import { OpenAPIHono } from '@hono/zod-openapi';
-import { zValidator } from '@hono/zod-validator';
-import { getLogger } from '@logtape/logtape';
-import type { Context } from 'hono';
-import { convertIntakeHandler } from '@/modules/practice-client-intakes/handlers/convert.handler';
-import { listIntakesHandler } from '@/modules/practice-client-intakes/handlers/list.handler';
+import * as handlers from '@/modules/practice-client-intakes/handlers';
 import * as routes from '@/modules/practice-client-intakes/routes';
-import { practiceClientIntakesService } from '@/modules/practice-client-intakes/services/practice-client-intakes.service';
-import {
-  intakeValidations,
-} from '@/modules/practice-client-intakes/validations/practice-client-intakes.validation';
-import type { BetterAuthInstance } from '@/shared/auth/better-auth';
-import { createBetterAuthInstance } from '@/shared/auth/better-auth';
-import { db } from '@/shared/database';
+import { createHonoApp } from '@/shared/router/factory';
 import { registerOpenApiRoutes } from '@/shared/router/openapi-docs';
-import type { Session } from '@/shared/types/BetterAuth';
-import type { AppContext, AppRouteHandler } from '@/shared/types/hono';
-import { sanitizeHeaders } from '@/shared/utils/logging';
-import { response } from '@/shared/utils/responseUtils';
 
-const app = new OpenAPIHono<AppContext>();
-const logger = getLogger(['practice-client-intakes', 'http']);
+const practiceClientIntakesApp = createHonoApp();
 
-const getOptionalSessionUserId = async function getOptionalSessionUserId(
-  c: Context<AppContext>,
-): Promise<string | undefined> {
-  const existingUserId: string | undefined = c.get('userId') ?? undefined;
-  if (existingUserId) {
-    return existingUserId;
-  }
-
-  try {
-    const auth: BetterAuthInstance = createBetterAuthInstance(db);
-    const session: Session | null | undefined = await auth.api.getSession({
-      headers: c.req.raw.headers,
-    });
-    return session?.user?.id ?? undefined;
-  } catch (err) {
-    logger.error('Failed to resolve optional session user id', {
-      error: err,
-      existingUserId,
-      headers: sanitizeHeaders(Object.fromEntries(c.req.raw.headers.entries())),
-    });
-    return undefined;
-  }
-};
-
-// GET /:slug/intake
-// Public intake page - returns organization details and payment settings
-app.get('/:slug/intake', zValidator('param', intakeValidations.slugParamSchema), async (c) => {
-  const { slug } = c.req.valid('param');
-  const result = await practiceClientIntakesService.getPracticeClientIntakeSettings(slug);
-  return response.fromResult(c, result);
-});
-
-
-// POST /create
-// Creates payment intent for practice client intake
-// Will be mounted at /api/practice/client-intakes/create
-app.post('/create', zValidator('json', intakeValidations.createPracticeClientIntakeSchema), async (c) => {
-  const body = c.req.valid('json');
-  const clientIp = c.req.header('x-forwarded-for')
-    || c.req.header('cf-connecting-ip')
-    || c.req.header('x-real-ip');
-  const userAgent = c.req.header('user-agent');
-
-  // Get user_id from authenticated session context (if available)
-  // Never trust user_id from request body to prevent spoofing
-  const sessionUserId = await getOptionalSessionUserId(c);
-
-  const result = await practiceClientIntakesService.createPracticeClientIntake({
-    ...body,
-    // Override any user_id from request body with session user_id
-    user_id: sessionUserId ?? undefined,
-    clientIp,
-    userAgent,
-    origin: c.req.header('origin'),
-  });
-
-  return response.fromResult(c, result, 201);
-});
-
-// POST /:uuid/checkout-session
-// Creates Stripe Checkout Session for existing intake
-app.post(
-  '/:uuid/checkout-session',
-  zValidator('param', intakeValidations.uuidParamSchema),
-  async (c) => {
-    const { uuid } = c.req.valid('param');
-    const sessionUserId = await getOptionalSessionUserId(c);
-    const result = await practiceClientIntakesService.createPracticeClientIntakeCheckoutSession({
-      uuid,
-      user_id: sessionUserId,
-      origin: c.req.header('origin'),
-    });
-    return response.fromResult(c, result, 201);
-  },
+// ==================== PRACTICE CLIENT INTAKES ====================
+practiceClientIntakesApp.openapi(routes.getIntakeSettingsRoute, handlers.getIntakeSettingsHandler);
+practiceClientIntakesApp.openapi(routes.createPracticeClientIntakeRoute, handlers.createPracticeClientIntakeHandler);
+practiceClientIntakesApp.openapi(
+  routes.createPracticeClientIntakeCheckoutSessionRoute,
+  handlers.createPracticeClientIntakeCheckoutSessionHandler,
 );
+practiceClientIntakesApp.openapi(routes.updatePracticeClientIntakeRoute, handlers.updatePracticeClientIntakeHandler);
+practiceClientIntakesApp.openapi(routes.getPracticeClientIntakeStatusRoute, handlers.getPracticeClientIntakeStatusHandler);
+practiceClientIntakesApp.openapi(routes.getPracticeClientIntakePostPayStatusRoute, handlers.getPracticeClientIntakePostPayStatusHandler);
+practiceClientIntakesApp.openapi(routes.claimPracticeClientIntakeRoute, handlers.claimPracticeClientIntakeHandler);
+practiceClientIntakesApp.openapi(routes.triggerIntakeInvitationRoute, handlers.triggerIntakeInvitationHandler);
+practiceClientIntakesApp.openapi(routes.listIntakesRoute, handlers.listIntakesHandler);
+practiceClientIntakesApp.openapi(routes.convertIntakeRoute, handlers.convertIntakeHandler);
 
+registerOpenApiRoutes(practiceClientIntakesApp, routes);
 
-// PUT /:uuid
-app.put(
-  '/:uuid',
-  zValidator('param', intakeValidations.uuidParamSchema),
-  zValidator('json', intakeValidations.updatePracticeClientIntakeSchema),
-  async (c) => {
-    const { uuid } = c.req.valid('param');
-    const { amount } = c.req.valid('json');
-
-    const result = await practiceClientIntakesService.updatePracticeClientIntake(uuid, amount);
-    return response.fromResult(c, result);
-  },
-);
-
-
-// GET /:uuid/status
-// Gets payment status
-// Will be mounted at /api/practice/client-intakes/:uuid/status
-app.get('/:uuid/status', zValidator('param', intakeValidations.uuidParamSchema), async (c) => {
-  const { uuid } = c.req.valid('param');
-  const sessionUserId = c.get('userId') ?? undefined;
-  const result = await practiceClientIntakesService.getPracticeClientIntakeStatus(uuid, sessionUserId);
-  return response.fromResult(c, result);
-});
-
-// GET /post-pay/status
-app.get(
-  '/post-pay/status',
-  zValidator('query', intakeValidations.checkoutSessionStatusQuerySchema),
-  async (c) => {
-    const { session_id } = c.req.valid('query');
-    const result = await practiceClientIntakesService.getPracticeClientIntakePostPayStatus(session_id);
-    return response.fromResult(c, result);
-  },
-);
-
-// POST /claim
-app.post(
-  '/claim',
-  zValidator('json', intakeValidations.claimPracticeClientIntakeSchema),
-  async (c) => {
-    const { session_id } = c.req.valid('json');
-    const sessionUserId = c.get('userId');
-
-    if (!sessionUserId) {
-      return response.unauthorized(c, 'Authentication required to claim intake');
-    }
-
-    const result = await practiceClientIntakesService.claimPracticeClientIntakePayment({
-      session_id,
-      user_id: sessionUserId,
-    });
-    return response.fromResult(c, result);
-  },
-);
-
-// POST /:uuid/invite
-// Triggers an invitation for the user associated with this intake
-const triggerIntakeInvitationHandler: AppRouteHandler<typeof routes.triggerIntakeInvitationRoute> = async (c) => {
-  const { uuid } = c.req.valid('param');
-  const sessionUserId = c.get('userId');
-
-  if (!sessionUserId) {
-    return response.unauthorized(c, 'Authentication required to trigger invitation');
-  }
-
-  const result = await practiceClientIntakesService.triggerIntakeInvitation(
-    uuid,
-    sessionUserId,
-    c.req.raw.headers,
-  );
-
-  return response.fromResult(c, result);
-};
-
-app.openapi(routes.triggerIntakeInvitationRoute, triggerIntakeInvitationHandler);
-
-app.openapi(routes.listIntakesRoute, listIntakesHandler);
-app.openapi(routes.convertIntakeRoute, convertIntakeHandler);
-
-registerOpenApiRoutes(app, routes);
-
-export default app;
+export default practiceClientIntakesApp;
