+++ src/modules/practice-client-intakes/index.ts
@@ -1,3 +1,15 @@
-import practiceClientIntakesApp from '@/modules/practice-client-intakes/http';
+/**
+ * Practice Client Intakes Module
+ *
+ * Main entry point for the practice client intakes module
+ */
+
+import practiceClientIntakesApp from './http';
 
 export default practiceClientIntakesApp;
+
+// Export types
+export * from './types/practice-client-intakes.types';
+
+// Export schemas for migrations
+export * from './database/schema/practice-client-intakes.schema';
